hefan

+++ datastruct.h
@@ -108,7 +108,7 @@ typedef struct element_sel
 {   
 	gdouble pri_axis_ape;
 	gdouble sec_axis_ape;
-	gdouble	primary_axis_s;
+	gint	primary_axis_s;
 }ELEMENT_SEL ,*ELEMENT_SEL_P;
